+++ src/screens/app/Detail/styles.js
@@ -16,13 +16,12 @@ export default StyleSheet.create({
     paddingHorizontal: resize(12),
     backgroundColor: colors.backgroundColorApp,
   },
-  categoryName: {
+  title: {
     fontFamily: fonts.montserratSemiBold,
     fontSize: resize(18),
     textTransform: 'uppercase',
     color: colors.titleCategoryColor,
-    paddingBottom: resize(30),
-    paddingTop: resize(10),
-    backgroundColor: colors.backgroundColorApp
+    paddingTop: resize(15),
+    paddingBottom: resize(20),
   }
 });